random commit backup, fuck that shit, 4 frames and it crashed FUCK THAT SHIT IM OUT≈

+++ Student_Cuda_Image/src/cpp/core/02_Mandelbrot_Julia/moo/host/Fractal.h
@@ -22,7 +22,7 @@ class Fractal: public AnimableFonctionel_I
 
     public:
 
-	Fractal(int w, int h, float dt, int nMin,int nMax,bool julia,double cx,double cy,double xMin, double xMax, double yMin, double yMax);
+	Fractal(int w, int h, int dt, int nMin,int nMax,bool julia,double cx,double cy,double xMin, double xMax, double yMin, double yMax);
 	virtual ~Fractal(void);
 
 	/*--------------------------------------*\
